requirements.txt
flask>=3.0
python-docx>=1.1
mammoth>=1.8
lxml>=5.0
requests>=2.31
